+++ src/styles/global.tsx
@@ -1,12 +1,11 @@
 import { createGlobalStyle } from 'styled-components';
+import { Theme } from '@styles/theme';
 
 export const GlobalStyle = createGlobalStyle`
   body {
     font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
     color: #222;
     margin: 0;
-    background: #DFEFFF;
-
-    svg { width: inherit; }
+    background: ${Theme.secondary};
   }
 `;