+++ src/styles/theme.tsx
@@ -1,3 +1,5 @@
 export const Theme = {
+  gray: '#AAAAAA',
   primary: '#254463',
+  secondary: '#DFEFFF',
 };